refactor: moved script function inside utils.ts and reordered the code a bit, then called getLocationByCoords() inside index.tsx
